+++ src/main/java/ch/uzh/ifi/hase/soprafs24/exceptions/OpenAIClientException.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs24.service;
+package ch.uzh.ifi.hase.soprafs24.exceptions;
 
 public class OpenAIClientException extends RuntimeException {
     public OpenAIClientException(String message, Throwable cause) {
