+++ src/main/java/ch/uzh/ifi/hase/soprafs24/utils/GameStatisticsUtil.java
@@ -1,4 +1,4 @@
-package ch.uzh.ifi.hase.soprafs24.service;
+package ch.uzh.ifi.hase.soprafs24.utils;
 
 import ch.uzh.ifi.hase.soprafs24.entity.User;
 import ch.uzh.ifi.hase.soprafs24.game.result.Outcome;
